requirements.txt
torch==2.0.0
fastapi==0.95.0
uvicorn==0.22.0
Pillow==9.5.0
